requirements.txt
adbutils==0.8.3
aircv==1.4.6
allure-pytest==2.8.31
allure-python-commons==2.8.31
anytree==2.8.0
apkutils2==1.0.0
appdirs==1.4.4
asgiref==3.3.1
attrdict==2.0.1
attrs==20.3.0
atx==1.2.0
AxmlParserPY==0.0.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==19.10b0
bunch==1.0.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cigam==0.0.3
click==7.1.2
colorama==0.4.4
colorlog==4.7.2
cryptography==3.3.1
cssselect==1.1.0
cssutils==1.0.2
ddt==1.4.1
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
Django==3.1.4
et-xmlfile==1.0.1
facebook-wda==1.3.2
filetype==1.0.7
futures==3.0.5
gevent==21.1.2
greenlet==1.0.0
httprunner==3.1.4
humanize==3.2.0
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
ipaddr==2.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.9.5
logger==1.4
loguru==0.4.1
logzero==1.5.0
lxml==4.6.2
maproxy==0.0.12
MarkupSafe==1.1.1
more-itertools==8.7.0
numpy==1.19.5
opencv-python==4.5.1.48
openpyxl==3.0.6
packaging==20.8
paramiko==2.7.2
pathspec==0.8.1
Pillow==8.1.0
pluggy==0.13.1
premailer==3.7.0
progress==1.5
psutil==5.8.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyelftools==0.27
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
pytz==2020.4
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
rethinkdb==2.4.2.post1
retry==0.9.2
rpm==0.0.2
selenium==3.141.0
sentry-sdk==0.14.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
stepspy==1.4.0
tinydb==4.3.0
toml==0.10.2
tornado==6.1
tqdm==4.5.0
typed-ast==1.4.2
uiautomator==1.0.2
uiautomator2==2.12.1
urllib3==1.26.2
wcwidth==0.2.5
weditor==0.6.3
whichcraft==0.6.1
wrapt==1.12.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
yagmail==0.14.245
zope.event==4.5.0
zope.interface==5.2.0
